+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # 2.x branch
 ## 2.11 branch
-### 2.11.5
+### 2.12.0
+* removed global `firstValue()`. Use instead `CollectionInterface::first()`;
+* removed global `isLocalhost()`. Use instead the `is('localhost')` detector;
+* removed global `getClientIp()`. Use instead `ServerRequest::clientIp()`;
 * updated for CakePHP 3.4.
 
 ### 2.11.4